Move functions for creating/extracting tarballs into utils
Useful for other code paths which will handle zstd compression and
decompression.
